requirements.txt
faster-whisper==1.0.3
httpx==0.27.0
anthropic==0.39.0
python-dotenv==1.0.1
pyaudio
openwakeword
ddgs
Pillow
pynput
psutil
numpy
requests
